PolyCircle beginnings
Added most of the code for PolyCircle collision detection.
Only missing the looping of points and finding the MinMax points for the Poly.
Should be a breeze once this is sorted out.
Another note, shoving the functions to get min/max poitns and projection range into the respective Colliders. i.e. GetMinMax() / Project()